pyproject.toml
[project]
name = "llm-telegram-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[job-queue]~=21.3",
    "pydantic~=2.7",
    "pypdf~=5.5",
    "typed-argument-parser~=1.10",
    "pyyaml~=6.0",
    "telegramify-markdown~=0.1",
    "openai~=1.54",
    "google-api-python-client~=2.145",
    "readability-lxml~=0.8",
    "lxml[html_clean]",                     # for readability-lxml
    "html2text~=2024.2",
    "msgpack~=1.0",
    "wolframalpha~=5.1",
    "mistune~=3.0",
    "expecttest~=0.2",
    "telethon~=1.37",
    "langchain-community~=0.3",
    "litellm~=1.74",
]
